requirements.txt
django==1.9.8
djangorestframework==3.4.0
numpy==1.11.1
psycopg2==2.6.2
pyowm==2.3.2